pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "diskcache>=5.6.3",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyinstaller>=6.18.0",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "python-pptx>=1.0.2",
    "uvicorn>=0.40.0",
]
